rename AccountPropertyAddressTransactionBuffer to  AccountPropertiesAddressTransactionBuffer
